requirements.txt
absl_py==1.4.0
keras==2.9.0
matplotlib==3.5.3
memory_profiler==0.61.0
numpy==1.21.6
pandas==1.3.5
scikit_learn==1.0.2
tensorflow_gpu==2.9.0
torch==1.13.1
torchdata==0.5.1
torchtext==0.14.1
tqdm==4.64.1
